+++ CHANGELOG.md
@@ -3,12 +3,13 @@ Changelog
 
   * Changed album name loading, now it reads the real album name, if available from Jango.
     Issue: tiliado/nuvola-app-jango#12
+  * Integrate volume management. Issue: tiliado/nuvola-app-jango#16
 
 2.3 - February 12th, 2017
 -------------------------
 
   * Ported to Nuvola SDK.
-  
+
 2.2 - August 27th, 2016
 -----------------------